pyproject.toml
[project]
name = "songranker-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "supabase>=2.10.0",
    "rapidfuzz>=3.11.0",
    "redis>=7.1.0",
    "rq>=2.6.1",
    "gunicorn>=23.0.0",
    "slowapi>=0.1.9",
    "playwright>=1.57.0",
    "jinja2>=3.1.6",
    "numpy>=2.4.1",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "locust>=2.33.0",
]
